Books debugged
-Click to leave open was broken because you need to click the outside to make it work. TRapping you INSIDE THE BOOK!!!!

LMAOO, so thats why we need the toggle button.

I still left the hover function because its fun as  fuck to watch those pages flip on desktop :]

Toggle button brough in new bug that only allowed to me to use the first toggle because of the swipper but the updated script fixes this! ;D